backend/requirements.txt
Flask>=2.0
flask-cors>=3.0
python-dotenv>=1.0
openai>=1.3.7